requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
curl_cffi==0.11.1
cycler==0.12.1
fonttools==4.58.0
frozendict==2.4.6
hmmlearn==0.3.3
idna==3.10
iniconfig==2.1.0
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.3
multitasking==0.0.11
numpy==2.2.6
packaging==25.0
pandas==2.2.3
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
protobuf==6.31.0
pycparser==2.22
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
soupsieve==2.7
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websockets==15.0.1
yfinance==0.2.61
